\name{CVine-class}
\docType{class}

\alias{CVine-class}
\alias{dvine,CVine-method}
\alias{pitVine,CVine-method}
\alias{pvine,CVine-method}
\alias{rvine,CVine-method}
\alias{show,CVine-method}

\title{Class for Canonical Vines}
\description{
  Extends the \code{\linkS4class{RVine}} class.
}
\section{Slots}{
  See \code{\linkS4class{RVine}} for information about the inherited slots.
}
\section{Methods}{
  \describe{
    \item{show}{\code{signature(object = "CVine")}:}
    \item{dvine}{\code{signature(vine = "CVine")}:}
    \item{pvine}{\code{signature(vine = "CVine")}:}
    \item{rvine}{\code{signature(vine = "CVine")}:}
    \item{pitVine}{\code{signature(vine = "CVine")}:}
  }
}
\seealso{
  \code{\linkS4class{Vine}}, 
  \code{\linkS4class{RVine}}, 
  \code{\linkS4class{DVine}}.
}
\keyword{classes}
\keyword{multivariate}
\keyword{distribution}
\keyword{models}
