\name{dimnames-methods}
\docType{methods}

\alias{dimnames-methods}
\alias{dimnames,Vine-method}

\title{Methods to Retrieve the Names of the Dimensions of a Vine}
\section{Methods}{
\describe{

\item{\code{signature(x = "Vine")}}{
  \code{\link{dimnames}} for \code{\linkS4class{Vine}}.
}
}}
\keyword{methods}
