\name{DVine-class}
\docType{class}

\alias{DVine-class}
\alias{dvine,DVine-method}
\alias{pitVine,DVine-method}
\alias{pvine,DVine-method}
\alias{rvine,DVine-method}
\alias{show,DVine-method}

\title{Class for D-vines}
\description{
  Extends the \code{\linkS4class{RVine}} class.
}
\section{Slots}{
  See \code{\linkS4class{RVine}} for information about the inherited slots.
}
\section{Methods}{
  \describe{
    \item{show}{\code{signature(object = "DVine")}:}
    \item{dvine}{\code{signature(vine = "DVine")}:}
    \item{pvine}{\code{signature(vine = "DVine")}:}
    \item{rvine}{\code{signature(vine = "DVine")}:}
    \item{pitVine}{\code{signature(vine = "DVine")}:}
  }
}
\seealso{
  \code{\linkS4class{Vine}}, 
  \code{\linkS4class{RVine}}, 
  \code{\linkS4class{CVine}}.
}
\keyword{classes}
\keyword{multivariate}
\keyword{distribution}
\keyword{models}
