\name{dvine-methods}
\docType{methods}

\alias{dvine-methods}
\alias{dvine,CVine-method}
\alias{dvine,DVine-method}

\title{Vine Density Evaluation Methods}
\section{Methods}{
\describe{

\item{\code{signature(vine = "CVine")}}{
  Density function evaluation for \code{\linkS4class{CVine}}.
}

\item{\code{signature(vine = "DVine")}}{
  Density function evaluation for \code{\linkS4class{DVine}}.
}
}}
\keyword{methods}
\keyword{distribution}
