\name{gofVine-class}
\docType{class}

\alias{gofVine-class}
\alias{show,gofVine-method}

\title{Class for the Results of Vine Goodness-of-fit Tests}
\description{
}
\section{Slots}{
  \describe{
    \item{\code{method}:}{Object of class \code{"character"}}
    \item{\code{pvalue}:}{Object of class \code{"numeric"}}
    \item{\code{statistic}:}{Object of class \code{"numeric"}}
  }
}
\section{Methods}{
  \describe{
    \item{show}{\code{signature(object = "gofVine")}:}
  }
}
\keyword{classes}
\keyword{multivariate}
\keyword{models}
\keyword{htest}
