\name{pitVine-methods}
\docType{methods}

\alias{pitVine-methods}
\alias{pitVine,CVine-method}
\alias{pitVine,DVine-method}

\title{Vine Probability Integral Transformation Methods}
\section{Methods}{
\describe{

\item{\code{signature(vine = "CVine")}}{
  PIT for \code{\linkS4class{CVine}}.
}

\item{\code{signature(vine = "DVine")}}{
  PIT for \code{\linkS4class{DVine}}.
}
}}
\keyword{methods}
