\name{RVine-class}

\docType{class}

\alias{RVine-class}

\title{Class for Regular Vines}

\description{
Extends the \code{\linkS4class{Vine}} class to represent the Regular vines.
Currently general regular vines are not implemented.
}

\section{Slots}{
See \code{\linkS4class{Vine}} for information about the inherited slots.
}

\seealso{
\code{\linkS4class{Vine}},
\code{\linkS4class{CVine}},
\code{\linkS4class{DVine}}.
}

\keyword{classes}
\keyword{multivariate}
\keyword{models}
