\name{rvine-methods}
\docType{methods}

\alias{rvine-methods}
\alias{rvine,CVine-method}
\alias{rvine,DVine-method}

\title{Vine Random Number Generation Methods}
\section{Methods}{
\describe{

\item{\code{signature(vine = "CVine")}}{
  Random number generation for \code{\linkS4class{CVine}}.
}

\item{\code{signature(vine = "DVine")}}{
  Random number generation for \code{\linkS4class{DVine}}.
}
}}
\keyword{methods}
\keyword{distribution}
