\name{vines-package}
\docType{package}
\alias{vines-package}
\alias{vines}
\title{Multivariate Dependence Modeling with Vines}
\description{
  Vines are graphical models for pair-copula constructions
  that allows building high-dimensional distributions from bivariate
  copulas and marginal density functions. This package contains S4
  classes for vines (especially Canonical vines and D-vines) and
  methods for inference, goodness-of-fit tests, density evaluation,
  distribution function evaluation and simulation.
}
\references{
  Bedford, T. and R. M. Cooke (2001). Probability density decomposition 
  for conditionally dependent random variables modeled by vines. 
  \emph{Annals of Mathematics and Artificial Intelligence} 32, 245-268.

  Bedford, T. and R. M. Cooke (2002). Vines - a new graphical model for 
  dependent random variables. \emph{Annals of Statistics} 30, 1031-1068.

  Joe, H. (1996). Families of \eqn{m}{m}-variate distributions with given 
  margins and \eqn{m(m-1)/2}{m(m-1)/2} bivariate dependence parameters.
  In L. Ruschendorf, B. Schweizer and M. D. Taylor (Eds.), 
  \emph{Distributions with Fixed Marginals and Related Topics}.

  Aas, K., C. Czado, A. Frigessi, and H. Bakken (2009). Pair-copula
  constructions of multiple dependence. \emph{Insurance: Mathematics
  and Economics} 44, 182-198.
}
\keyword{package}
\keyword{multivariate}
\keyword{models}
