\name{CVine-class}

\docType{class}

\alias{CVine-class}
\alias{show,CVine-method}

\title{Class for C-vines}

\description{
Extends the \code{\linkS4class{RVine}} class to represent C-vines.
}

\section{Slots}{
See \code{\linkS4class{RVine}} for information about the inherited slots.
}

\section{Methods}{
\describe{
   \item{show}{\code{signature(object = "CVine")}: 
     Print a text representation of the C-vine.}
}
}

\seealso{
\code{\linkS4class{Vine}},
\code{\linkS4class{RVine}},
\code{\linkS4class{DVine}}.
}

\keyword{classes}
\keyword{multivariate}
\keyword{models}
