\name{DVine-class}

\docType{class}

\alias{DVine-class}
\alias{show,DVine-method}

\title{Class for D-vines}

\description{
Extends the \code{\linkS4class{RVine}} class to represent D-vines.
}

\section{Slots}{
See \code{\linkS4class{RVine}} for information about the inherited slots.
}

\section{Methods}{
\describe{
   \item{show}{\code{signature(object = "DVine")}: 
     Print a text representation of the D-vine.}
}
}

\seealso{
\code{\linkS4class{Vine}},
\code{\linkS4class{RVine}},
\code{\linkS4class{CVine}}.
}

\keyword{classes}
\keyword{multivariate}
\keyword{models}
