\name{fitVine-class}
\docType{class}

\alias{fitVine-class}
\alias{show,fitVine-method}

\title{Class for the Results of Vine Inference}
\description{
}
\section{Slots}{
  \describe{
    \item{\code{vine}:}{Object of class \code{"Vine"}}
    \item{\code{observations}:}{Object of class \code{"numeric"}}
    \item{\code{method}:}{Object of class \code{"character"}}
  }
}
\section{Methods}{
  \describe{
    \item{show}{\code{signature(object = "fitVine")}:}
  }
}
\seealso{
  \code{\linkS4class{fitVineML}}.
}
\keyword{classes}
\keyword{multivariate}
\keyword{models}
