\name{h-methods}
\docType{methods}

\alias{h-methods}
\alias{h,claytonCopula-method}
\alias{h,copula-method}
\alias{h,gumbelCopula-method}
\alias{h,indepCopula-method}
\alias{h,normalCopula-method}
\alias{h,tCopula-method}

\title{Methods for the h-functions}
\description{
}
\section{Methods}{
\describe{

\item{\code{signature(copula = "copula")}}{
}

\item{\code{signature(copula = "indepCopula")}}{
}

\item{\code{signature(copula = "normalCopula")}}{
}

\item{\code{signature(copula = "tCopula")}}{
}

\item{\code{signature(copula = "claytonCopula")}}{
}

\item{\code{signature(copula = "gumbelCopula")}}{
}
}}
\keyword{methods}
