\name{hinverse-methods}
\docType{methods}

\alias{hinverse-methods}
\alias{hinverse,claytonCopula-method}
\alias{hinverse,copula-method}
\alias{hinverse,indepCopula-method}
\alias{hinverse,normalCopula-method}
\alias{hinverse,tCopula-method}

\title{Methods for the Inverse of the h-functions}
\description{
}
\section{Methods}{
\describe{

\item{\code{signature(copula = "copula")}}{
}

\item{\code{signature(copula = "indepCopula")}}{
}

\item{\code{signature(copula = "normalCopula")}}{
}

\item{\code{signature(copula = "tCopula")}}{
}

\item{\code{signature(copula = "claytonCopula")}}{
}
}}
\keyword{methods}
