\name{RVine-class}
\docType{class}

\alias{RVine-class}

\title{Class for Regular Vines}
\description{
}
\section{Slots}{
  See \code{\linkS4class{Vine}} for information about the inherited slots.
}
\seealso{
  \code{\linkS4class{Vine}}, 
  \code{\linkS4class{CVine}}, 
  \code{\linkS4class{DVine}}.
}
\keyword{classes}
\keyword{multivariate}
\keyword{distribution}
\keyword{models}
