\name{vines-package}

\docType{package}

\alias{vines-package}
\alias{vines}

\title{Multivariate Dependence Modeling with Vines}

\description{
Vines are graphical models for pair-copula constructions that allows building
high-dimensional distributions from bivariate copulas and marginal density 
functions. This package contains S4 classes for vines (especially Canonical 
vines and D-vines) and methods for inference, goodness-of-fit tests, density 
evaluation, distribution function evaluation and simulation.
}

\references{
Aas, K. and Czado, C. and Frigessi, A. and Bakken, H. (2009)
Pair-copula constructions of multiple dependence.
\emph{Insurance: Mathematics and Economics} \bold{44}, 182--198.

Bedford, T. and Cooke, R. M. (2001) Probability density decomposition 
for conditionally dependent random variables modeled by vines. 
\emph{Annals of Mathematics and Artificial Intelligence} \bold{32},
245--268.

Bedford, T. and Cooke, R. M. (2002) Vines -- a new graphical model for
dependent random variables. \emph{Annals of Statistics} \bold{30},
1031--1068.

Joe, H. (1996) Families of \eqn{m}{m}-variate distributions with given
margins and \eqn{m(m-1)/2}{m(m-1)/2} bivariate dependence parameters.
In Ruschendorf, L. and Schweizer, B. and Taylor, M. D. (Eds.),
\emph{Distributions with fixed marginals and related topics}.
}

\keyword{package}
\keyword{models}
\keyword{multivariate}
